+++ Moq.AutoMocker.Generators.Tests/TestGeneratorTests.cs
@@ -2,10 +2,9 @@
 using Microsoft.CodeAnalysis.Testing;
 using Microsoft.CodeAnalysis.Text;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+using VerifyCS = Moq.AutoMocker.Generators.Tests.CSharpSourceGeneratorVerifier<Moq.AutoMocker.Generators.UnitTestSourceGenerator>;
 
-using VerifyCS = Moq.AutoMocker.TestGenerator.Tests.CSharpSourceGeneratorVerifier<Moq.AutoMocker.TestGenerator.UnitTestSourceGenerator>;
-
-namespace Moq.AutoMocker.TestGenerator.Tests;
+namespace Moq.AutoMocker.Generators.Tests;
 
 [TestClass]
 public class TestGeneratorTests
@@ -136,7 +135,7 @@ public interface ILogger<Controller> { }
                     GetSourceFile(expected, "ControllerTests.g.cs")
                 }
             }
-            
+
         }.RunAsync();
     }
 
@@ -522,6 +521,6 @@ public class Controller
 
     private static (string FileName, SourceText SourceText) GetSourceFile(string content, string fileName)
     {
-        return (Path.Combine("Moq.AutoMocker.TestGenerator", "Moq.AutoMocker.TestGenerator.UnitTestSourceGenerator", fileName), SourceText.From(content, Encoding.UTF8));
+        return (Path.Combine("Moq.AutoMocker.Generators", "Moq.AutoMocker.Generators.UnitTestSourceGenerator", fileName), SourceText.From(content, Encoding.UTF8));
     }
 }
