+++ Moq.AutoMocker.Generators/SyntaxReceiver.cs
@@ -2,6 +2,8 @@ using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 
+using static Moq.AutoMocker.Generators.AutoMock;
+
 namespace Moq.AutoMocker.Generators;
 
 public class SyntaxReceiver : ISyntaxContextReceiver
@@ -21,29 +23,35 @@ public class SyntaxReceiver : ISyntaxContextReceiver
             : null;
     }
 
+    private bool GetSkipNullableParameters(AttributeSyntax attributeSyntax)
+    {
+        return attributeSyntax.ArgumentList?.Arguments.Count > 0 &&
+            attributeSyntax.ArgumentList!.Arguments
+                .Select(x => x.Expression)
+                .OfType<MemberAccessExpressionSyntax>()
+                .FirstOrDefault() is { Name.Identifier.ValueText: IgnoreNullableParametersEnumValue };
+    }
+
+
     public void OnVisitSyntaxNode(GeneratorSyntaxContext context)
     {
         if (context.Node is ClassDeclarationSyntax classDeclaration &&
             context.SemanticModel.GetDeclaredSymbol(classDeclaration) is INamedTypeSymbol symbol &&
             classDeclaration.AttributeLists.SelectMany(x => x.Attributes)
-                .Select(a =>
-                {
-                    if (context.SemanticModel.GetTypeInfo(a).Type?.Name == AutoMock.ConstructorTestsAttribute)
-                    {
-                        if (GetTargetType(a) is { } targetType &&
-                            context.SemanticModel.GetTypeInfo(targetType).Type is INamedTypeSymbol sutType)
-                        {
-                            return sutType;
-                        }
-                        Diagnostic diagnostic = Diagnostics.MustSpecifyTargetType.Create(a.GetLocation(),
-                            symbol.OriginalDefinition.ToDisplayString(SymbolDisplayFormat.CSharpErrorMessageFormat));
-                        DiagnosticMessages.Add(diagnostic);
-                    }
-                    return null;
-                })
-                .FirstOrDefault(a => a is not null) is { } sutType
-            )
+                .Select(a => context.SemanticModel.GetTypeInfo(a).Type?.Name == ConstructorTestsAttribute ? a : null)
+                .FirstOrDefault(a => a is not null) is { } attribute)
         {
+            if (!(GetTargetType(attribute) is { } targetType) ||
+                context.SemanticModel.GetTypeInfo(targetType).Type is not INamedTypeSymbol sutType)
+            {
+                Diagnostic diagnostic = Diagnostics.MustSpecifyTargetType.Create(attribute.GetLocation(),
+                    symbol.OriginalDefinition.ToDisplayString(SymbolDisplayFormat.CSharpErrorMessageFormat));
+                DiagnosticMessages.Add(diagnostic);
+                return;
+            }
+
+            bool skipNullableParameters = GetSkipNullableParameters(attribute);
+
             if (!classDeclaration.Modifiers.Any(x => x.IsKind(SyntaxKind.PartialKeyword)))
             {
                 Diagnostic diagnostic = Diagnostics.TestClassesMustBePartial.Create(classDeclaration.GetLocation(),
@@ -66,7 +74,6 @@ public class SyntaxReceiver : ISyntaxContextReceiver
                 int nullIndex = 0;
                 foreach (IParameterSymbol parameter in ctor.Parameters)
                 {
-                    if (parameter.Type.IsValueType) continue;
                     sut.NullConstructorParameterTests.Add(new NullConstructorParameterTest()
                     {
                         Parameters = parameters,
@@ -81,7 +88,8 @@ public class SyntaxReceiver : ISyntaxContextReceiver
             {
                 Namespace = namespaceDeclaration,
                 TestClassName = testClassName,
-                Sut = sut
+                Sut = sut,
+                SkipNullableParameters = skipNullableParameters
             };
 
             TestClasses.Add(targetClass);
